refactor(cpp): 🚚 reduce TypeAliases.h to type aliases only
TypeAliases.h had grown into a grab-bag: an allocator, the row-store
accessor overloads, a sentinel constant, and -- below the closing
namespace -- three includes plus a forward declaration to break the
resulting cycle.

That tail existed for one reason: the OperatorIndex row accessors live in
monoprop:: while their callers live in monoprop::detail, so ADL cannot
find them and they must be declared before any template that calls them
is parsed. RowAccess.h now owns that ordering contract and states it, so
the forward declaration and all three trailing includes are gone.

Consumers that had been getting OperatorIndex/InvertedIndex/MPOperator
and core/Monomial.h transitively through that tail now include them
directly.

kMissingIndex moves to its only caller, layer_build/Common.h; that also
makes the existing OperatorIndex.h comment ("the operator store must not
depend on evolution headers") literally true rather than aspirational.

No functional changes.
