+++ cpp/monoprop/DefaultInitAllocator.h
@@ -1,0 +1,49 @@
+// Copyright 2026 Algorithmiq
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <memory>
+#include <new>
+#include <type_traits>
+#include <utility>
+
+namespace monoprop {
+
+// Allocator that default-initializes (no zero-fill) on the no-arg construct. Use only for buffers whose
+// every grown element is overwritten before read — otherwise it exposes indeterminate values.
+template <typename T, typename A = std::allocator<T>>
+struct default_init_allocator : A {
+    using a_traits = std::allocator_traits<A>;
+    template <typename U>
+    struct rebind {
+        using other = default_init_allocator<U, typename a_traits::template rebind_alloc<U>>;
+    };
+    using A::A;
+    default_init_allocator() = default;
+    template <typename U>
+    default_init_allocator(const default_init_allocator<U, typename a_traits::template rebind_alloc<U>> &o) noexcept
+        : A(static_cast<const typename a_traits::template rebind_alloc<U> &>(o)) {}
+
+    template <typename U>
+    void construct(U *ptr) noexcept(std::is_nothrow_default_constructible_v<U>) {
+        ::new (static_cast<void *>(ptr)) U;
+    }
+    template <typename U, typename... Args>
+    void construct(U *ptr, Args &&...args) {
+        a_traits::construct(static_cast<A &>(*this), ptr, std::forward<Args>(args)...);
+    }
+};
+
+} // namespace monoprop
