+++ cpp/monoprop/detail/operator/CMakeLists.txt
@@ -7,4 +7,5 @@ target_sources(
         "InvertedIndex.h"
         "MPOperator.h"
         "OperatorIndex.h"
+        "RowAccess.h"
 )
