+++ cpp/tests/majorana_cutoff_tests.cpp
@@ -23,6 +23,7 @@
 
 #include "monoprop/TypeAliases.h"
 #include "monoprop/algebra/MajoranaAlgebra.h"
+#include "monoprop/core/Monomial.h"
 
 using namespace monoprop;
 using cd = std::complex<double>;
